# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: omu-rest
  description: 'Laravel製RESTful APIサーバー。著者: [connect0459](https://github.com/connect0459)'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/api'
    description: ローカルホストのAPIエンドポイント
  -
    url: 'https://example.studyfortwo.org/omu-rest/public/api'
    description: lolipopのAPIエンドポイント
paths:
...
  /types_receives:
    get:
      tags:
        - types_receives
      summary: types_receivesのリストを取得する
      operationId: 0417ea5e25bfb8fd8ee49856f17016fa
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TypeReceive'
    post:
      tags:
        - types_receives
      summary: 新しいtypes_receivesを作成する
      operationId: 50090e82cc0eacb3d7b613a9ff868b17
      requestBody:
        description: TypeReceiveのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TypeReceive'
      responses:
        201:
          description: リソースが作成されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/TypeReceive'
  '/types_receives/{id}':
    get:
      tags:
        - types_receives
      summary: 特定のIDでtypes_receivesを取得する
      operationId: f84e4922d0eeb8c8d92bcbf48af06a74
      parameters:
        -
          name: id
          in: path
          description: types_receivesのID
          required: true
          schema:
            type: string
      responses:
        200:
          description: 成功した応答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeReceive'
        404:
          description: リソースが見つかりません
    put:
      tags:
        - types_receives
      summary: 特定のIDでtypes_receivesを更新する
      operationId: c1543df8694acb50827b80eb08cd5b4a
      parameters:
        -
          name: id
          in: path
          description: types_receivesのID
          required: true
          schema:
            type: string
      requestBody:
        description: TypeReceiveのデータ
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TypeReceive'
      responses:
        200:
          description: リソースが更新されました
          content:
            application/json:
              schema:
                type: object
                allOf:
                  -
                    $ref: '#/components/schemas/TypeReceive'
        404:
          description: リソースが見つかりません
    delete:
      tags:
        - types_receives
      summary: 特定のIDでtypes_receivesを削除する
      operationId: b1a600a99289da180209079339415e49
      parameters:
        -
          name: id
          in: path
          description: types_receivesのID
          required: true
          schema:
            type: string
      responses:
        204:
          description: リソースが削除されました
        404:
          description: リソースが見つかりません
components:
  schemas:
...
    TypeReceive:
      description: 'TypeReceive Model'
      properties:
        id:
          description: ID
          type: integer
          format: int64
          example: '1'
        type:
          description: 受取り方法
          type: string
          format: string
          example: 露店
        created_at:
          description: レコード作成日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
        updated_at:
          description: レコード更新日
          type: string
          format: date-time
          example: '2023-06-20T21:20:14.000000Z'
      type: object
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: api_token
      in: header